==> 99_main_biofire.R <==
# main biofire script
# source each file in turn

source("01_libraries_func_biofire.R")
source('02_excel_cleanup_biofire.R')
source('03_lab_biofire.R')
source("04_cdr_biofire.R")
source('05_cdtox_biofire.R')
source('06_merge_biofire.R')
source('11_table1_biofire.R')
source('12_infections_bar_biofire.R')
source('13_infagents_bar_biofire.R')
source('14_prognostic_biofire.R')
